added separate test file for esearchquery
